+++ packages/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @heathmont/moon-react
 
+## 0.9.0
+
+### Minor Changes
+
+- feat: update Authenticator, Breadcrumb, Dialog, Drawer, Dropdown, IconButton, Menu, Pagination, TabList, Snackbar, Carousel
+
 ## 0.8.0
 
 ### Minor Changes